move loadbarbutton to datasessions, removed unused css, changed console error to use alert store
